requirements.txt
flask>=3.1.0
pandas>=2.3.0
pymupdf4llm>=0.3.4
pymupdf>=1.23.0
google-genai>=1.65.0